+++ src/components/product/ProductCard/ProductCard.tsx
@@ -5,7 +5,7 @@ import {
   CardImage as UICardImage,
 } from '@faststore/ui'
 import { graphql, Link } from 'gatsby'
-import React, { memo, useMemo } from 'react'
+import React, { memo } from 'react'
 import { Badge, DiscountBadge } from 'src/components/ui/Badge'
 import { Image } from 'src/components/ui/Image'
 import Price from 'src/components/ui/Price'
@@ -22,7 +22,6 @@ interface Props {
   product: ProductSummary_ProductFragment
   index: number
   bordered?: boolean
-  outOfStock?: boolean
   variant?: Variant
   aspectRatio?: number
   buyButton?: ReactNode
@@ -43,34 +42,20 @@ function ProductCard({
   variant = 'vertical',
   bordered = false,
   aspectRatio = 1,
-  outOfStock = false,
   buyButton,
   ...otherProps
 }: Props) {
   const {
     isVariantOf: { name },
     image: [img],
-    offers: { lowPrice: spotPrice, offers },
+    offers: {
+      lowPrice: spotPrice,
+      offers: [{ listPrice, availability }],
+    },
   } = product
 
-  // TODO: Move this computation to the backend
-  const selectedOffer = useMemo(() => {
-    const lowestPriceOffer = offers.findIndex((x) => x.price === spotPrice)
-
-    if (lowestPriceOffer === -1) {
-      console.error(
-        'Could not find the lowest price product offer. Showing the first offer provided.'
-      )
-
-      return 0
-    }
-
-    return lowestPriceOffer
-  }, [spotPrice, offers])
-
-  const { listPrice } = offers[selectedOffer]
-
-  const linkProps = useProductLink({ product, selectedOffer, index })
+  const linkProps = useProductLink({ product, selectedOffer: 0, index })
+  const outOfStock = availability !== 'https://schema.org/InStock'
 
   return (
     <UICard